Add non-mpi catalyst example, other fixes
