+++ src/examples/CMakeLists.txt
@@ -26,7 +26,7 @@ if(BABELFLOW_FOUND)
 endif()
 
 # install using and python examples
-install(DIRECTORY using-with-cmake using-with-make python tutorial 
+install(DIRECTORY ascent-catalyst using-with-cmake using-with-make python tutorial 
         DESTINATION examples/ascent)
 
 if(MPI_FOUND)
