+++ src/examples/ascent-catalyst-mpi/ascent_actions.yaml
@@ -1,0 +1,9 @@
+-
+  action: "add_scenes"
+  scenes:
+    s1:
+      plots:
+        p1:
+          type: "pseudocolor"
+          field: "braid"
+      image_name: "out_ascent_catalyst_mpi_example"
